+++ lib/service/function.go
@@ -1,6 +1,8 @@
 package service
 
 import (
+	"errors"
+
 	"github.com/substra/orchestrator/lib/asset"
 	"github.com/substra/orchestrator/lib/common"
 	orcerrors "github.com/substra/orchestrator/lib/errors"
@@ -67,13 +69,14 @@ func (s *FunctionService) RegisterFunction(a *asset.NewFunction, owner string) (
 		Key:          a.Key,
 		Name:         a.Name,
 		Description:  a.Description,
-		Function:     a.Function,
+		Archive:      a.Archive,
 		Metadata:     a.Metadata,
 		Owner:        owner,
 		CreationDate: timestamppb.New(s.GetTimeService().GetTransactionTime()),
 		Inputs:       a.Inputs,
 		Outputs:      a.Outputs,
 		Status:       asset.FunctionStatus_FUNCTION_STATUS_WAITING,
+		Image:        &asset.Addressable{StorageAddress: "", Checksum: ""},
 	}
 
 	function.Permissions, err = s.GetPermissionService().CreatePermissions(owner, a.NewPermissions)
@@ -128,7 +131,7 @@ func (s *FunctionService) FunctionExists(key string) (bool, error) {
 	return s.GetFunctionDBAL().FunctionExists(key)
 }
 
-// UpdateFunction updates mutable fields of an function. List of mutable fields : name.
+// UpdateFunction updates mutable fields of an function. List of mutable fields : name, image.
 func (s *FunctionService) UpdateFunction(a *asset.UpdateFunctionParam, requester string) error {
 	s.GetLogger().Debug().Str("requester", requester).Interface("functionUpdate", a).Msg("Updating function")
 	err := a.Validate()
@@ -147,8 +150,17 @@ func (s *FunctionService) UpdateFunction(a *asset.UpdateFunctionParam, requester
 		return orcerrors.NewPermissionDenied("requester does not own the function")
 	}
 
-	// Update function name
+	// Update function name. Name cannot be blank.
 	function.Name = a.Name
+
+	// Update Image if given
+	if len(a.Image.GetChecksum()) != 0 {
+		function.Image = a.Image
+	}
+
+	if len(function.Image.GetChecksum()) == 0 && function.Status == asset.FunctionStatus_FUNCTION_STATUS_READY {
+		return orcerrors.FromValidationError(asset.FunctionKind, errors.New("Image checksum should not be empty when function status is ready"))
+	}
 
 	event := &asset.Event{
 		EventKind: asset.EventKind_EVENT_ASSET_UPDATED,